requirements.txt
requests
beautifulsoup4
openai
flask
flask-cors
supabase
werkzeug
python-dotenv